requirements.txt
# requirements.txt
pymupdf
sentence-transformers
numpy
flask
chromadb
cohere
python-dotenv